Rename packages for AR and BarcodeScan modules
