+++ modules/augmented-reality/src/main/java/com/gluonhq/attach/augmentedreality/impl/IOSAugmentedRealityService.java
@@ -25,9 +25,9 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.gluonhq.attach.ar.impl;
+package com.gluonhq.attach.augmentedreality.impl;
 
-import com.gluonhq.attach.ar.ARModel;
+import com.gluonhq.attach.augmentedreality.ARModel;
 import javafx.application.Platform;
 import javafx.beans.property.ReadOnlyBooleanProperty;
 import javafx.beans.property.ReadOnlyBooleanWrapper;
