+++ modules/augmented-reality/src/main/java/module-info.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, Gluon
+ * Copyright (c) 2019, 2020, Gluon
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -31,6 +31,6 @@ module com.gluonhq.attach.augmentedreality {
     requires com.gluonhq.attach.util;
     requires com.gluonhq.attach.storage;
 
-    exports com.gluonhq.attach.ar;
-    exports com.gluonhq.attach.ar.impl to com.gluonhq.attach.util;
+    exports com.gluonhq.attach.augmentedreality;
+    exports com.gluonhq.attach.augmentedreality.impl to com.gluonhq.attach.util;
 }
