--- modules/barcode-scan/src/main/java/com/gluonhq/attach/barcode/impl/DummyBarcodeScanService.java
@@ -1,6 +1,0 @@
-package com.gluonhq.attach.barcode.impl;
-
-import com.gluonhq.attach.barcode.BarcodeScanService;
-
-public abstract class DummyBarcodeScanService implements BarcodeScanService {
-}
