+++ modules/barcode-scan/src/main/java/com/gluonhq/attach/barcodescan/BarcodeScanService.java
@@ -25,7 +25,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.gluonhq.attach.barcode;
+package com.gluonhq.attach.barcodescan;
 
 import com.gluonhq.attach.util.Services;
