+++ modules/barcode-scan/src/main/java/com/gluonhq/attach/barcodescan/impl/AndroidBarcodeScanService.java
@@ -25,9 +25,9 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.gluonhq.attach.barcode.impl;
+package com.gluonhq.attach.barcodescan.impl;
 
-import com.gluonhq.attach.barcode.BarcodeScanService;
+import com.gluonhq.attach.barcodescan.BarcodeScanService;
 import javafx.application.Platform;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
