+++ modules/barcode-scan/src/main/java/com/gluonhq/attach/barcodescan/impl/DummyBarcodeScanService.java
@@ -1,0 +1,6 @@
+package com.gluonhq.attach.barcodescan.impl;
+
+import com.gluonhq.attach.barcodescan.BarcodeScanService;
+
+public abstract class DummyBarcodeScanService implements BarcodeScanService {
+}
